modified import in test_itertree_examples
