--- tests/test_lib.py
@@ -1,30 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-This code is taken from the itertree package:
-  _ _____ _____ _____ _____ _____ _____ _____
- | |_   _|   __| __  |_   _| __  |   __|   __|
- |-| | | |   __|    -| | | |    -|   __|   __|
- |_| |_| |_____|__|__| |_| |__|__|_____|_____|
-
-https://pypi.org/project/itertree/
-GIT Home:
-https://github.com/BR1py/itertree
-The documentation can be found here:
-https://itertree.readthedocs.io/en/latest/index.html
-
-The code is published under MIT license
-For more information see: https://en.wikipedia.org/wiki/MIT_License
-
-CONTENT DESCRIPTION:
-
-This files contains global helper methods used in the tests
-"""
-import importlib
-import sys
-
-def import_from_path(module_name, file_path):
-    spec = importlib.util.spec_from_file_location(module_name, file_path)
-    module = importlib.util.module_from_spec(spec)
-    sys.modules[module_name] = module
-    spec.loader.exec_module(module)
-    return module
